Possible Timer.Delay fix using CommandGroup based Programming
