+++ src/org/usfirst/frc4930/Zoot/commands/Shoot.java
@@ -1,33 +1,14 @@
 package org.usfirst.frc4930.Zoot.commands;
 
-import edu.wpi.first.wpilibj.Timer;
-import edu.wpi.first.wpilibj.command.Command;
-import org.usfirst.frc4930.Zoot.Robot;
+import edu.wpi.first.wpilibj.command.CommandGroup;
 
 /**
- * Shoot - shoots the boulder
+ *
  */
-public class Shoot extends Command {
-
-  protected void initialize() {}
-
-  protected void execute() {
-    Robot.intakeMotors.move(-1.0);
-    Timer.delay(0.75);
-    Robot.roller.move(1.0);
-    Timer.delay(0.1);
-  }
-
-  protected boolean isFinished() {
-    return true;
-  }
-
-  protected void end() {
-    Robot.intakeMotors.stop();
-    Robot.roller.stop();
-  }
-
-  protected void interrupted() {
-    end();
-  }
+public class Shoot extends CommandGroup {
+    
+    public  Shoot() {
+    	addSequential(new SpinUpShooter());
+    	addSequential(new SpinUpRoller());
+    }
 }
